Update README file with project details and instructions
This commit updates the README file with information about the Fitness-Landing-Page project, including its features, instructions for running the project locally, and information about the deployment. It also includes details about the technologies used in the project. This update will help new users understand the project better and be able to set it up easily.